+++ nitrite/src/main/java/org/dizitart/no2/transaction/DefaultTransactionalRepository.java
@@ -197,13 +197,13 @@ class DefaultTransactionalRepository<T> implements ObjectRepository<T> {
     }
 
     @Override
-    public void subscribe(CollectionEventListener listener) {
-        backingCollection.subscribe(listener);
+    public String subscribe(CollectionEventListener listener) {
+        return backingCollection.subscribe(listener);
     }
 
     @Override
-    public void unsubscribe(CollectionEventListener listener) {
-        backingCollection.unsubscribe(listener);
+    public void unsubscribe(String subscription) {
+        backingCollection.unsubscribe(subscription);
     }
 
     @Override